requirements_conda.txt
astropy
dill
numdifftools
numpy
pandas
python-lal
python-lalsimulation
scipy
sympy
tqdm
